--- apps/nextjs/src/components/features/menus/index.ts
@@ -1,12 +1,0 @@
-export * from './assign-project-owner-menu';
-export * from './assign-task-menu';
-export * from './assignee-icon-menu';
-export * from './assignee-menu';
-export * from './attachment-menu';
-export * from './editor-emoji-menu';
-export * from './editor-mention-menu';
-export * from './invite-collaborator-menu';
-export * from './invite-project-teammate-menu';
-export * from './menu-select';
-export * from './project-menu';
-export * from './tag-menu';
