+++ apps/nextjs/src/components/features/modals/share-project-modal/body/permission-menu/permission-menu.tsx
@@ -3,7 +3,7 @@ import {
   MenuSelect,
   MenuSelectList,
   MenuSelectTrigger,
-} from '@/components/features/menus';
+} from '@/components/features/menus/menu-select';
 import { Button } from '@/components/ui/button';
 import { Flex } from '@/components/ui/flex';
 import { Icon } from '@/components/ui/icon';
