+++ apps/nextjs/src/components/features/task-detail/task-detail-body/form/assignee/input.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import { memo, useCallback, useState } from 'react';
-import { AssigneeMenu } from '@/components/features/menus';
+import { AssigneeMenu } from '@/components/features/menus/assignee-menu';
 import { Input as AtomsInput } from '@/components/ui/input';
 import { useClickOutside } from '@/hooks/use-click-outside';
 import { useDisclosure } from '@/shared/chakra';
