+++ apps/nextjs/src/components/features/task-detail/task-detail-body/form/attachment/new-button.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { AttachmentMenu } from '@/components/features/menus';
+import { AttachmentMenu } from '@/components/features/menus/attachment-menu';
 import { Menu } from '@/components/ui/menu';
 import { NewBox } from '@/components/ui/new-box';
