+++ apps/nextjs/src/components/features/task-detail/task-detail-body/form/projects/input.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import { memo, useCallback, useState } from 'react';
-import { ProjectMenu } from '@/components/features/menus';
+import { ProjectMenu } from '@/components/features/menus/project-menu';
 import { Flex } from '@/components/ui/flex';
 import { Input as AtomsInput, type InputProps } from '@/components/ui/input';
 import { useClickOutside } from '@/hooks/use-click-outside';
