+++ apps/nextjs/src/components/features/tasks/tasks-header/sort-menu/sort-menu.tsx
@@ -2,7 +2,7 @@ import {
   MenuSelect,
   MenuSelectList,
   MenuSelectTrigger,
-} from '@/components/features/menus';
+} from '@/components/features/menus/menu-select';
 import { Button } from '@/components/ui/button';
 import { Icon } from '@/components/ui/icon';
 import { Menu } from '@/components/ui/menu';
