+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-assignee/input.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import { memo, useCallback, useState } from 'react';
-import { AssignTaskMenu } from '@/components/features/menus';
+import { AssignTaskMenu } from '@/components/features/menus/assign-task-menu';
 import { Input as AtomsInput } from '@/components/ui/input';
 import { useClickOutside } from '@/hooks/use-click-outside';
 import { useDisclosure } from '@/shared/chakra';
