+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-name/move-tasks-between-sections/menu-list/menu-list.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { MenuSelectList } from '@/components/features/menus';
+import { MenuSelectList } from '@/components/features/menus/menu-select';
 import {
   type TaskSection,
   useTasksTaskSectionByTaskId,
