+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-name/move-tasks-between-sections/move-tasks-between-sections.tsx
@@ -1,5 +1,8 @@
 import { memo, useCallback } from 'react';
-import { MenuSelect, MenuSelectTrigger } from '@/components/features/menus';
+import {
+  MenuSelect,
+  MenuSelectTrigger,
+} from '@/components/features/menus/menu-select';
 import { useTasksTask } from '@/components/features/tasks/hooks';
 import { Icon } from '@/components/ui/icon';
 import { IconButton } from '@/components/ui/icon-button';
