+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-projects/input/input.tsx
@@ -1,7 +1,7 @@
 import type React from 'react';
 import { memo, useCallback, useState } from 'react';
 import { ProjectChip } from '@/components/features/chips/project-chip';
-import { ProjectMenu } from '@/components/features/menus';
+import { ProjectMenu } from '@/components/features/menus/project-menu';
 import { Flex } from '@/components/ui/flex';
 import { Input as AtomsInput } from '@/components/ui/input';
 import { Wrap, WrapItem } from '@/components/ui/wrap';
